+++ apps/web/src/main.tsx
@@ -13,6 +13,7 @@ import "./ui-preferences.css";
 import "./kidung-ux.css";
 import "./direct-manipulation.css";
 import "./kidung-touch-targets.css";
+import "./reading-surfaces.css";
 import "./kidung-responsive.css";
 import "./calm-liturgical.css";